Changed structure of auth
Changed structure of auth
