--- zakat/accounts/users_auth/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class UsersAuthConfig(AppConfig):
-    name = 'users_auth'
