--- zakat/accounts/users_auth/urls.py
@@ -1,27 +1,0 @@
-"""zakat.news URL Configuration
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/3.0/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
-from django.urls import path,include
-from .views import *
-from django.conf import settings
-
-
-app_name = 'users_auth'
-
-urlpatterns = [
-    path('', include('social_django.urls',namespace='social')),
-    path('logout/', user_logout,name='logout'),
-    path('test',testView),
-]
